LSMTree/requirements.txt
python-snappy==0.7.2
bitarray==2.9.2
